Scroll to view when quoted message is updated
