+++ src/script/components/message/ContentMessage/index.tsx
@@ -17,7 +17,7 @@
  *
  */
 
-import React from 'react';
+import React, {useEffect} from 'react';
 
 import {QualifiedId} from '@wireapp/api-client/src/user';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
@@ -65,6 +65,7 @@ export interface TextMessageProps extends Omit<MessageActions, 'onClickResetSess
   isLastDeliveredMessage: boolean;
   message: ContentMessage;
   onClickButton?: (message: ContentMessage, assetId: string) => void;
+  onContentUpdated?: () => void;
   quotedMessage?: ContentMessage;
   selfId: QualifiedId;
   shouldShowAvatar: boolean;
@@ -161,6 +162,7 @@ const TextMessage: React.FC<TextMessageProps> = ({
   onClickMessage,
   onClickLikes,
   onClickButton,
+  onContentUpdated,
   onLike,
 }) => {
   const {headerSenderName, timestamp, ephemeral_caption, ephemeral_status, assets, other_likes} =
@@ -172,6 +174,22 @@ const TextMessage: React.FC<TextMessageProps> = ({
       'assets',
       'other_likes',
     ]);
+
+  useEffect(() => {
+    if (message.hasAssetText()) {
+      // add a listener to any changes to the assets. This will warn the parent that the message has changed
+      const assetSubscription = message.assets.subscribe(onContentUpdated);
+      // also listen for link previews on a single Text entity
+      const previewSubscription = (message.getFirstAsset() as Text).previews.subscribe(onContentUpdated);
+      return () => {
+        if (assetSubscription) {
+          assetSubscription.dispose();
+          previewSubscription.dispose();
+        }
+      };
+    }
+    return undefined;
+  }, []);
 
   const avatarSection = shouldShowAvatar ? (
     <div className="message-header">
@@ -227,6 +245,7 @@ const TextMessage: React.FC<TextMessageProps> = ({
       {message.quote() && (
         <MessageQuote
           conversation={conversation}
+          onContentUpdated={onContentUpdated}
           quote={message.quote()}
           selfId={selfId}
           findMessage={findMessage}
